backend/requirements.txt
fastapi==0.121.2
uvicorn==0.35.0
pydantic==2.12.3
pydantic_core==2.41.4
httpx==0.27.2
starlette==0.49.3
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.5
protobuf==5.29.5
grpcio==1.76.0
grpcio-status==1.71.2
requests==2.32.4
certifi==2025.7.9
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
tqdm==4.67.1
python-dotenv==1.2.1
typing-extensions==4.15.0
python-multipart
Pillow
